Link all movies with their genres in a separate table and update database for full functionality

+++ MCIA/MCIA/include/MovieGenre.h
@@ -5,17 +5,17 @@ class MovieGenre
 {
 public:
 	MovieGenre() = default;
-	MovieGenre(const uint16_t movieId, const uint16_t genreId);
+	MovieGenre(const uint32_t movieId, const uint16_t genreId);
 
-	uint16_t GetMovieId() const;
+	uint32_t GetMovieId() const;
 	uint16_t GetGenreId() const;
 
-	void SetMovieId(const uint16_t movieId);
+	void SetMovieId(const uint32_t movieId);
 	void SetGenreId(const uint16_t genreId);
 
 	void LinkGenreMovie();
 private:
-	uint16_t m_movieId;
+	uint32_t m_movieId;
 	uint16_t m_genreId;
 };
